requirements.txt
boto3==1.35.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.6.0
structlog==23.2.0
pytest==7.4.3
black==23.11.0